pyproject.toml
[tool.poetry]
name = "Hearthbot"
version = "0.1.0"
description = "Just another QQ bot for Hearthstone"
license = "AGPL-3.0-or-later"
authors = ["ZelKnow <shenghuang@stu.pku.edu.cn>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = {extras = ["websockets", "fastapi", "httpx"], version = "^2.3.3"}
hearthstone = "^5.40.4"
Pillow = "^9.0.0"
nonebot-adapter-onebot = "^2.4.4"
nonebot-plugin-apscheduler = "^0.5.0"
nb-cli = "^1.4.2"
nonebot-adapter-qq = "^1.5.1"

[tool.poetry.dev-dependencies]
nonebot-plugin-test = "^0.3.0"
flake8 = "^3.9.2"
yapf = "^0.31.0"
toml = "^0.10.2"

[nonebot.plugins]
plugins = []
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.core.masonry.api"
